new port.. it syncs your IMAP folder to your ~/Maildir/
written by the mutt dude :)

isync is a command line utility for synchronizing a remote IMAP mailbox
with a local maildir-style mailbox. This is useful for working in
disconnected mode, such as on a laptop. Modifications made locally and
remotely are synchronized so that no message status flags are lost.

Bug:
Submitted by:
Reviewed by:
Approved by:
Obtained from:


git-svn-id: http://svn.macports.org/repository/macports/trunk@6174 d073be05-634f-4543-b044-5fe20cf6d1d6