Bug: Submitted by: Reviewed by: Approved by: Obtained from:
Fix a bug in the way a file is being given to the filemap in register_bulk_files.  This would only affect direct-mode users, but would prevent any file tracking after a port was installed, because the file entry in the filemap would contain all of the file information (MD5, etc), as stored in the receipt, rather than just the filename.  Thanks to rshaw@ for catching this.


git-svn-id: http://svn.macports.org/repository/macports/trunk@8481 d073be05-634f-4543-b044-5fe20cf6d1d6