Bug:            #2426 Submitted by:   mww@ Reviewed by:    dluke@geeklair.net, waqar@, rshaw@, michaelm@
Feature Addition: "StartupItem creation"

this adds new keywords to port(1), to automatically create an appropriate
startupitem for the OS.
The current state only has code for Darwin-specific StartupItems, but the code
is prepared to create e. g. BSD or Linux startup scripts on those platforms.

new keywords are:
	startupitem.create [yes|no]
	startupitem.requires [name of other startup item]
	startupitem.name [the name of the startup item, default=${name}]
	startupitem.start [executable to call to start the daemon]
	startupitem.stop [executable to call to stop the daemon]
	startupitem.restart executable to call to restart the daemon]

the default for startupitem.{start|stop|restart} is to call
	${prefix}/etc/startup/${name}.sh {start|stop|restart}


git-svn-id: http://svn.macports.org/repository/macports/trunk@10468 d073be05-634f-4543-b044-5fe20cf6d1d6