Critical concurrent bug fix.
filemap close would revert to the previous version of the filemap if it was
open read/only. This is a critical bug when running concurrent instances of
port(1) and could yield to incorrect filemap representation of what is
installed.

The easiest way to display this bug is to select 4 or 5 ports and run:
sudo port deactivate port{1,2,3,4,5} && sudo port activate port{1,2,3,4,5}
in 5 different terminal windows. One port instance will succeed (however, I'm
not sure the data will be correct) and (some of) the four other instances will
fail.

The commit includes the fix and a test case (the test fails with the previous
version of the pextlib).


git-svn-id: http://svn.macports.org/repository/macports/trunk@13359 d073be05-634f-4543-b044-5fe20cf6d1d6