Submitted by:   jmpp@
Small updates to the Portfile and ReadMe.rtf file to get the appropriate
OS info in it depending on where the pkg is built.
Also added some explanatory notes to the Portfile and switched the home URL
to our brand new www.darwinports.org address, which we now fully control, woot!


git-svn-id: http://svn.macports.org/repository/macports/trunk@15319 d073be05-634f-4543-b044-5fe20cf6d1d6