Submitted by:   jmpp@
Portfile:

General improvements to the long_description so we can reinplace it better on
the pkg target to display more accurate OS info. Improve Darwin version reinplace.


ReadMe.rtf:

Note "Port Authority" may not be installed on the local drive (provide the URL
where it can be downloaded from), as it is a separate install. Now there's a
disclaimer clearly stating "Port Authority" is a third party product and not
officiall supported by the DarwinPorts project.


git-svn-id: http://svn.macports.org/repository/macports/trunk@15439 d073be05-634f-4543-b044-5fe20cf6d1d6