Port:		midgard-core Version:	1.7.4 Revision:	1 Bug:		7954 Bug-URL:	<http://bugzilla.opendarwin.org/show_bug.cgi?id=7954>
Update port www/midgard-core
(Verbatim from the bug)
This update helps to build other parts of Midgard. Mysql directories are linked
to build directories and are recorded wrong in midgard-config, so when later
other parts are handled, they can't be configured , because midgard-config is
asked about include and libs directories, which are not found.


git-svn-id: http://svn.macports.org/repository/macports/trunk@17171 d073be05-634f-4543-b044-5fe20cf6d1d6