Add an option to completely skip the checksum phase.
Setting checksum.skip=yes will skip the checksuming phase. This should never be used in a
portfile, but might be useful on the command line, especially while testing or
trying to install an old version of a port, or a version where the checksum is incorrect.
A better solution where the checksum is wrong would be to fix and report the bad checksum.

For instance, to install an outdated version of grep without hacking the portfile, do:

	port install grep version=2.5 checksum.skip=yes

Instead of skipping the checksum calculation, you could also explicitly specify the checksum:

	port install grep version=2.5 checksums="md5 thechecksum"

Warnings:

	- Overriding of the port version on the command line is for dire straights only.
	  Side effects are to be expected. Patches and other portfile tweaks may not
	  be appropriate for the specified version. Remember: you are causing a specific version
	  of the source to be built using Portfile rules for the version in the Portfile.
	  You will need to do a make clean following such an operation, as port doesn't
	  know which version of the port has been built.
	  You have been warned.

	- Any use of checksum.skip in a portfile is considered inappropriate.


git-svn-id: http://svn.macports.org/repository/macports/trunk@17214 d073be05-634f-4543-b044-5fe20cf6d1d6