Submitted by:   jmpp@ Reviewed by:    jberry@
Addition of a new "Project Membership" chapter to the online guide, as a response
to the points raised on the "What should I do?" thread on the main user list.

To all those reading this commit, please do feel free to read the new addition
and suggest improvements/fixes as you see fit, consider this just a first draft
(please do pay close attentio to xml syntax, as I only learned it as I typed along).


git-svn-id: http://svn.macports.org/repository/macports/trunk@17429 d073be05-634f-4543-b044-5fe20cf6d1d6