Bug:            7845 Submitted by:   rhwood@ Reviewed by: Approved by: Obtained from:
Bump revision to trigger outdated/upgradable status

Change port from a port that installs a package to a port that uninstalls itself and installs instead a different port (abiword-x11).

Given that the contents of abiword-x11 @2.4.4_0 are identical to the contents of abiword2 @2.4.4_0 there are probably some tricks I could engage in to make the switch between ports fast and painless, but this method has the advantage of being simple and reliable.


git-svn-id: http://svn.macports.org/repository/macports/trunk@18481 d073be05-634f-4543-b044-5fe20cf6d1d6