Add xar code into macports base build.
xar will be built and installed only if xar is not present in the base system, as it is rumored to be on Leopard.
This means that xar will be build and installed into /opt/local on Tiger, and that somebody who has already installed the xar port, or who decides to install the xar port on such a system, might run into a conflict. The number of such people is probably very low, according to bbraun.


git-svn-id: http://svn.macports.org/repository/macports/trunk@23687 d073be05-634f-4543-b044-5fe20cf6d1d6