gnome/gnotify: replace missing files from ${filespath} rather than port:automake
The current GNotify source has broken symlinks for files that it requires from
automake instead of the actual files.  port:gnotify currently fixes this by
copying them for the share directory of port:automake.  However, since GNotify
doesn't actually require automake to be run before building, we can just put the required files into ${filespath} and copy them over.  The files included here come from automake-1.10.

Of the files so added here, install-sh is under the MIT/X License, and
mkinstalldirs is in the public domain.  mkinstalldirs is under the GPL, but it
has an exception stating that, if it is included in a package that uses
autoconf, it may be distributed under that package's licence.  I may be
stretching this exception a bit by claiming that we are okay to copy
mkinstalldirs here because we use autoconf somewhere else in MacPorts.  If
anyone has any concerns, I can remove these files and restore the dependence on
port:automake, though I'll have to to do some hackery to retrieve the version of
the installed port:automake, as we have to get around the fact that
port:automake installs these files into ${prefix}/share/automake-${version}.

Ultimately, of course, the GNotify developers should fix the package so we don't
have to do this, so I will notify them of the problem.


git-svn-id: http://svn.macports.org/repository/macports/trunk@27400 d073be05-634f-4543-b044-5fe20cf6d1d6