Neat improvement over Chris' latest change to the layout of this page: downward alphabetically sorted columns of categories! Each <li> of the containing <ul> is one column, containing in turn a <ul> whose <li> entries are the items of the corresponding column. Adapt the XHTML markup for this, don't use a div for every single category but rather a "categories" encompasing one.
I used two explicit values to achieve this:

 * 15, the number of entries per column, which I think is OK (10 may be too little; 20 may be too much);
 * 5, the number of columns displaying across the entire page. This is a lame limitation which I used as a working shortcut to complete
   this commit. My eyes are currently too tired to figure out how to abstract this code to cover an arbitrary number of columns
   until the pool returned by 'mysql_fetch_assoc()' (querying the db for categories) is exhausted.

The two numbers happen to work together at the moment because 15 * 5 == 75, which is exactly the number of categories we currently have!
If that changes then I'll have to increase the number of allowed columns or we'll get an incomplete listing, so lets hope other eyes
(Chris and Ryan? ;-) ) can chime in and improve my technique.

Adapt the CSS to this new display, noting my inability to separate the W3C tags from the categories containing <ul>... helping eyes? ;-)


git-svn-id: http://svn.macports.org/repository/macports/trunk@31583 d073be05-634f-4543-b044-5fe20cf6d1d6