The ChangeLog was orignally meant to be a user-parsable file listing only major changes to base, so that we could use it in our release process to advertise our new goodies. Unfortunately, lately it has turned into a little bit of a file tracking commits to base, with a little bit of user level information and a lot of gaps in between to do either one right, bleh!
Therefore add a real user-parsable NEWS file in which new features and bug fixes for each release will be advertised,
starting with the contents of 1.6.0.

Here's hoping that from now on the ChangeLog will turn into a file that truly tracks commits to base, with proper revision
numbers for each entry and ticket numbers in case of bug fixes, after being spared from having to remain user-parsable.


git-svn-id: http://svn.macports.org/repository/macports/trunk@31640 d073be05-634f-4543-b044-5fe20cf6d1d6