Change portname to py25-gsl, as it should have been from the start.
To all committers: when using a given port as the basis of a new one, please always remember to make sure
you adapt the portname for the latter, as having duplicate portname entries in our tree is not allowed
(not only does it mess up informational operations, but it could also wreck havoc for a user that somehow
manages to install the two colliding ports).

Yes, I'm aware we don't have strong checks against this harmful situation within MacPorts' base, and I'm also
aware that we should... But, as we currently don't, let warnings like this raise committers' awareness of the
issue, while we develop a *real* solution.

Thanks!


git-svn-id: http://svn.macports.org/repository/macports/trunk@33525 d073be05-634f-4543-b044-5fe20cf6d1d6