Change the $by value of the dependency search back to 'library', as using 'dependency' broke clicking on a port's dependency's href and getting that sole port as a search result (I missed updating the relevant stuff there for things to work); the entire 4500+ ports list would be returned instead!
Also comment out searching by dependency for the time being, as the functionality is not exactly what I thought: searching for a dependency simply
returns a page for the port name, rather than the ports that depend on what as searched for (what I thought was the behavior). I'll code that behavior
later on and re-enable searching by dependency.


git-svn-id: http://svn.macports.org/repository/macports/trunk@35135 d073be05-634f-4543-b044-5fe20cf6d1d6