doc-new: New target "guide-chunked" which generates a chunked version of the guide.
The table of contents is automatically added to every page after the
generation and the section links are clickable like the normal guide. All
files are created in guide/html/chunked.

The tclsh binary has to be available as a simple Tcl script is used to add the
table of contents. If you know a better way to do this please change it.


git-svn-id: http://svn.macports.org/repository/macports/trunk@38342 d073be05-634f-4543-b044-5fe20cf6d1d6