base/aclocal.m4, base/configure: Do expansion on ~ when looking for the home directory of a user.
Most ports expect applications_dir to be an absolute path starting with a
slash. If this is not the case, constructs like ${destroot}${applications_dir}
will fail (work/destroot~foo/). This copies files outside the destroot so they
will not get installed.

Thanks to Kevin Reid (kpreid) for the report.


git-svn-id: http://svn.macports.org/repository/macports/trunk@48083 d073be05-634f-4543-b044-5fe20cf6d1d6