cdparanoia: Updated to version 10.2
I integrated the osx_interface changes that were in the osx version of paranoia (0.9.8) into 10.2 from upstream.

This version has but ignores the '-d' command line option since many rippers (eg abcde) expect that option.
*someone* should eventually add support for this option to choose which drive to rip from, but that someone is not me.

There is a ton of ifdef-fu based on __APPLE__ to determine if cdparanoia builds for linux or darwin.


git-svn-id: http://svn.macports.org/repository/macports/trunk@51693 d073be05-634f-4543-b044-5fe20cf6d1d6