The port fails in destroot with the +doc variant due to documentation build changes in 2.2.4.
Also, if the +doc variant is not specified then the documentation is
still installed.

The --docdir command line argument passed to configure is not honored
by docs/Makefile.in and it instead uses $(book_name) as the directory
name to install into.  Trying to set book_name by passing it on make's
command line argument causes to build to fail, so I took the approach
of installing everything which installs the extra docs into
${prefix}/share/doc/libsigc++-2.0 and then moving the contents into
${prefix}/share/doc/libsigc++-${version} if the variant it set.  Then
always delete libsigc++-2.0.


git-svn-id: http://svn.macports.org/repository/macports/trunk@56560 d073be05-634f-4543-b044-5fe20cf6d1d6