Fix build of libdlna.
Add inclusion of <stdlib.h> to upnp-dms.c.  Prior to Snow Leopard, this must
have been picked up implicitly via one of the other headers, but no longer.
This resulted in the compiler automatically generating an incorrect declaration
for malloc() and failing to compile this file.

Disable parallel build.  libdlna's "all" target has two dependencies: a "build"
target which builds the library, and a "test" target which verifies that the
library was built correctly.  Running these two targets in parallel is
unproductive.


git-svn-id: http://svn.macports.org/repository/macports/trunk@58734 d073be05-634f-4543-b044-5fe20cf6d1d6