Add support for setting cxx_stdlib in macports.conf
This will be added to configure.cxxflags -stdlib=${cxx_stdlib} when the C++
compiler is clang.  This is a developer option to aide in testing ports with
libc++.  Please do not file bug reports when using this option unless you
are also providing a fix for the problem being reported.

It is reccomended that you reinstall all ports from source when changing this
option (make sure you set buildfromsource always).


git-svn-id: http://svn.macports.org/repository/macports/trunk@110369 d073be05-634f-4543-b044-5fe20cf6d1d6