base: setup signal handling in mportinit
Given the code already in registry2.0/portimage.tcl this will finally prevent
inconsistent state when an activation operation is interrupted. The code there
will correctly catch the error and roll back the changes made to the
filesystem.

There might be further places where MacPorts could be smarter now that is has
signal handling code, but I think this change is already a huge improvement.


git-svn-id: http://svn.macports.org/repository/macports/trunk@118298 d073be05-634f-4543-b044-5fe20cf6d1d6