p5-mac-carbon: Prevent incorrect arch detection (#46411)
The build tries to determine the machine architecture using the running
Perl, adding an arch flag to LDDLFLAGS, LDFLAGS, and CCFLAGS. If the
running Perl is not i386 or PPC, the build gets it wrong, causing
compile and link errors.

And perl-1.0 wants to handle this anyway, so just let it.


git-svn-id: https://svn.macports.org/repository/macports/trunk@131237 d073be05-634f-4543-b044-5fe20cf6d1d6