python3{_select,3,4,5}: Don't link `python3-32`
See comment:39:ticket:34326.

The `python3.x-32` executables are only installed for universal builds,
and I can't think of a good way of structuring the `port select`
configuration so that things don't break when switching between
universal and non-universal.

Instead of working ourselves into contortions, let's just drop that link
and hope that no one wants it.


git-svn-id: https://svn.macports.org/repository/macports/trunk@136752 d073be05-634f-4543-b044-5fe20cf6d1d6