select-1.0: Rebuild clients affected by #47849
In r136723, I introduced a bug into select-1.0 (since fixed in r136794)
that caused client ports to execute their post-destroot, pre/post-install,
and pre/post-activate scripts with an incorrect `name`. Some of these
failures were fatal, but others were silent and produced subtly broken
installations, which need to be fixed with a revbump.


git-svn-id: https://svn.macports.org/repository/macports/trunk@136823 d073be05-634f-4543-b044-5fe20cf6d1d6