Add gnu_info group to run install-info on activate
GNU info requires running
  install-info $infofile $prefix/share/info/dir
on activation and
  install-info --delete $infofile $prefix/share/info/dir
on deactivation. Let's have a PortGroup to support that.

We should really have a generic trigger mechanism to get this right
rather than doing it manually for each port, especially since we don't
have a lint check that warns you if you forget it.


git-svn-id: https://svn.macports.org/repository/macports/trunk@140714 d073be05-634f-4543-b044-5fe20cf6d1d6